Fix the Windows Visual Studio Project so both Inklecate and ink-engind.dll references the same .NET version of Newtonsoft's JSON library
Fix Windows Visual Studio so that both Inklecate and ink-engine.dll
reference the same .NET version of the Newtonsoft JSON dll.
Inklecate was set to use Newtonsoft.JSON.dll for .NET 4.5 which is
incompatible with Unity3d. This was causing compilation exceptions to be
thrown when using the ink-unity-integration project for Windows. Note:
after this fixed version is built, its output should then be copied into
the ink-unity-integration project to fix it for the Windows users.