Convert ink-engine-runtime.csproj to a PCL (Portable Class Library)
This makes the DLL usable by a wider range of consumers. I'm using
profile 259 as that seems to have the broadest compatibility (see
http://danrigby.com/2014/04/16/xamarin-pcl-profile-notes/)

With this change, the NuGet package should work out of the box on
new installations of Xamarin, as well as VS 2013 and later.

It should not affect Unity users as they're including the source
directly.