Change NuGet target framework to "netstandard1.0"
This should make the DLL that we publish via NuGet as compatible
with as many different consumers as possible. This creates a
portable library that's usable from Mono, Xamarin, Windows Phone
and Windows.

Notably, it is *not* usable from Unity (which still uses an
older runtime based on .net 3.5) but Unity has a separate packaging
scheme and doesn't use NuGet.

Note: according to the MS docs, I should just need a single dependency
on "NETStandard.Library", but that doesn't work, apparently due to a
bug in NuGet (https://github.com/dotnet/corefx/issues/10445). I've
therefore used a workaround suggested in that bug report, and
explicitly listed the various System.* packages we depend on.