Need to copy storyState._originalCallstack
Not sure how this has not been an issue before, but it's recently started causing errors in our external function calls - perhaps they just hit a certain level of complexity? The issue seen in the inkjs prototype is that _originalCallstack is null by the time the external function finishes, when it then attempts to put that state back in again. The solution here fixes the issue in inkjs but I've not repo'd the issue in "real" ink.