Kill experimental lookups feature
Was an attempt to reduce the data size when saving in order to improve performance. Turned out to not be necessary after all since on Heaven’s Vault we found an alternative solution - saving on a separate thread while using a state patch internally to prevent state modifications while saving. Lookups feature adds bloat to the engine and isn’t so useful so doesn’t need to be kept around.