Correct spelling of "pedant"
The first change was ambiguous, "pendants" means something along "things that are equal to", whereas you meant "pedants", people who are very concerned about the accuracy of things (like I am). 
The second change: "pendantic" isn't a word. You meant to write "pedantic". 
I'm unsure whether this is a volitional typo or you didn't know about the correct spelling of pedant(ic), because it's kinda funny as it is.