Patch for joe!
InkList: Adds == operator for InkLists, and some other changes Jon(?) wrote for Overboard but never merged in
ListDefinitionsOrigin: Null input check. Can't recall why we added this!
Story: These were changes used in the InkPlayerEditorWindow
Path+StoryState: This is a new feature! It enables you to get a more practical path string for the story state, which I've found super helpful as a way of checking where the player REALLY is in the story (which can then be used to rewind them, track state, find bugs, etc)